+++ app/services/tests/navigate.test.js
@@ -1,43 +1,15 @@
-import {
-  setTopLevelNavigator,
-  navigatorObject,
-  navigate
-} from './NavigationService';
+import { navigatorObject, navigate } from '../NavigationService';
 import { set } from 'lodash';
 import { NavigationActions } from '@react-navigation/compat';
-
-jest.mock('lodash', () => ({
-  set: jest.fn()
-}));
-
-describe('setTopLevelNavigator', () => {
-  afterEach(() => {
-    jest.clearAllMocks();
-  });
-  it('sets the navigator object with the provided reference', () => {
-    const navigatorRef = { goBack: 'goBack' };
-    setTopLevelNavigator(navigatorRef);
-    expect(set).toHaveBeenCalledWith(
-      navigatorObject,
-      'navigator',
-      navigatorRef
-    );
-  });
-});
-
-const mockDispatch = jest.fn();
-
-// Mock the NavigationActions.navigate function
 jest.mock('@react-navigation/compat', () => ({
   NavigationActions: {
     navigate: jest.fn()
   }
 }));
-
+const mockDispatch = jest.fn();
 describe('navigate', () => {
   beforeEach(() => {
-    // eslint-disable-next-line fp/no-mutation
-    navigatorObject.navigator = { dispatch: mockDispatch };
+    set(navigatorObject, 'navigator', { dispatch: mockDispatch });
   });
 
   afterEach(() => {
